refactor to display help center link from UnknownError
