+++ app/javascript/packages/document-capture/components/unknown-error.tsx
@@ -1,12 +1,15 @@
 import type { ComponentProps } from 'react';
+import { useContext } from 'react';
 import { useI18n, HtmlTextWithStrongNoWrap } from '@18f/identity-react-i18n';
 import { FormStepError } from '@18f/identity-form-steps';
+import MarketingSiteContext from '../context/marketing-site';
 
 interface UnknownErrorProps extends ComponentProps<'p'> {
   unknownFieldErrors: FormStepError<{ front: string; back: string }>[];
   isFailedDocType: boolean;
   remainingAttempts: number;
   altFailedDocTypeMsg?: string | null;
+  hasDismissed: boolean;
 }
 
 function UnknownError({
@@ -14,8 +17,15 @@ function UnknownError({
   isFailedDocType = false,
   remainingAttempts,
   altFailedDocTypeMsg = null,
+  hasDismissed,
 }: UnknownErrorProps) {
   const { t } = useI18n();
+  const { getHelpCenterURL } = useContext(MarketingSiteContext);
+  const helpCenterLink = getHelpCenterURL({
+    category: 'verify-your-identity',
+    article: 'how-to-add-images-of-your-state-issued-id',
+    location: 'document_capture_review_issues',
+  });
   const errs =
     !!unknownFieldErrors &&
     unknownFieldErrors.filter((error) => !['front', 'back'].includes(error.field!));
@@ -33,11 +43,13 @@ function UnknownError({
       </p>
     );
   }
-  if (err) {
+  if (err && !hasDismissed) {
+    return <p key={err.message}>{err.message}</p>;
+  }
+  if (err && hasDismissed) {
     return (
       <p key={err.message}>
-        {err.message}
-        <a href={err.helpCenterLink}>{t('doc_auth.info.review_examples_of_photos')}</a>
+        {err.message} <a href={helpCenterLink}>{t('doc_auth.info.review_examples_of_photos')}</a>
       </p>
     );
   }
